Объекты json и простые объекты объединены в один файл
